refactored search
