+++ SkateSpotsMap/Views/Search/DirectionsButtonView.swift
@@ -1,0 +1,24 @@
+//
+//  DirectionsButtonView.swift
+//  SkateSpotsMap
+//
+//  Created by Dillon Shu on 12/10/23.
+//
+
+import SwiftUI
+
+struct DirectionsButtonView: View {
+    var body: some View {
+        Image(systemName: "arrow.triangle.turn.up.right.diamond.fill")
+            .resizable()
+            .aspectRatio(contentMode: .fit)
+            .frame(width: 40)
+            .foregroundStyle(.green)
+    }
+}
+
+struct DirectionsButtonView_Previews: PreviewProvider {
+    static var previews: some View {
+        DirectionsButtonView()
+    }
+}
