+++ SkateSpotsMap/Views/Search/SearchRowView.swift
@@ -18,35 +18,9 @@ struct SearchRowView: View {
                         Text(skateSpot.name)
                             .fontWeight(.bold)
                             .font(.title3)
-                        HStack(spacing: 4){
-                            var rating_string = String(format: "%.2f", skateSpot.rating_avg)
-                            Text(rating_string).multilineTextAlignment(.leading)
-                                .font(.system(size:10))
-                                .fontWeight(.light)
-                            Image(systemName: "star.fill")
-                                .resizable()
-                                .aspectRatio(contentMode: .fit)
-                                .scaledToFit()
-                                .frame(width: 12)
-                                .foregroundStyle(.yellow)
-                            //Text("(102)").multilineTextAlignment(.leading)
-                            Text("(" + String((skateSpot.reviews).count) + ")").multilineTextAlignment(.leading)
-                                .font(.system(size:10))
-                                .fontWeight(.light)
-                        }
-                        HStack{
-                            Text("Tags:").font(.system(size:14)).multilineTextAlignment(.leading)
-                            //ForEach(skateSpot.tags, id: \.self){ tag in
-                            ForEach(skateSpot.tags, id: \.self){ tag in
-                                Text(tag).fontWeight(.light).font(.system(size:10)).padding(.top, 2)
-                            }
-                        }
-                        HStack{
-                            Text("Features:").font(.system(size:14)).multilineTextAlignment(.leading)
-                            ForEach(skateSpot.features, id: \.self){ feature in
-                                Text(feature).fontWeight(.light).font(.system(size:10)).padding(.top, 2)
-                            }
-                        }
+                        SearchRatingView(skateSpot:skateSpot)
+                        SearchTagsView(skateSpot: skateSpot)
+                        SearchFeaturesView(skateSpot: skateSpot)
                     }
                     Spacer()
                     VStack {
@@ -58,11 +32,7 @@ struct SearchRowView: View {
                             let mapsURL = URL(string: "maps://?dirflg=w&saddr=&daddr=\(lat),\(long)")
                             UIApplication.shared.open(mapsURL!,options: [:],completionHandler: nil)
                         } label: {
-                            Image(systemName: "arrow.triangle.turn.up.right.diamond.fill")
-                                .resizable()
-                                .aspectRatio(contentMode: .fit)
-                                .frame(width: 40)
-                                .foregroundStyle(.green)
+                            DirectionsButtonView()
                         }.buttonStyle(.plain)
                     }
                 }
